Use template w/ mode to output the styles of nodes.
Instead of using choose-cases (if-then-else), process the nodes
with a mode `normal-node-with-style`. This allows for easy reuse of
these cases at several places.

Please note: According to <http://www.w3.org/TR/xslt.html#conflict> if
more than one template rule matches with same priority the one that
occurs *last* in the stylesheet has to be choosen. This iy why
@STYLE_REF is now before @LOCALIZED_STYLE_REF.