Add a Mindmap representing the current preferences structure.
The mindmap was generated using helper-xslt for building a mind-map
out of ´preferences.xml`, texts have been inserted from
translations/Resources_en.properties (using a script) and styls have
been added.

The result as generated by
 xsltproc freeplane/ant/mm2preferences.xsl \
    freeplane/resources/xml/preferences.mm \
    > freeplane/resources/xml/preferences.xml
has been verified to result in the same file as the current
`preferences.xml`.

Also add a build-rule to not include this file in the distribution.