Add a Mindmaps representing the current menu structures.
The mindmap was generated using helper-xslt for building a mind-map
out of ´*menu.xml`, texts have been inserted from
translations/Resources_en.properties (using a script) and styles have
been added.

The result as generated by
  xsltproc freeplane/ant/mm2menu.xsl \
    freeplane/resources/xml/$menu.mm \
    > freeplane/resources/xml/$menu.xml
(with $menu being one of mindmapmodemenu, filemodemenu, stylemodemenu)
has been verified to result in the same file as the current
`$menu.xml`.

Also add a build-rule to not include the mindmaps in the distribution.