Fix #2386 Special characters in add-on translations wrongly displayed
Added method escapeUtf8 to org.freeplane.core.util.TextUtils. Method
escapeUtf8 is a stripped down version StringEscapeUtils.escapeJava. It
only escapes UTF-8 non-ASCII characters. It escapes them to \uXXXX where
XXXX is the hex code of the escaped character. It solves this bug
because it does not escape backslashes and double quotes.

Replaced method "StringEscapeUtils.escapeJava" with
"TextUtils.escapeUtf8" on line 456 of
org.freeplane.main.addons.AddOnProperties.

As part of the fix of this bug the add-on devtools is also changed to
use TextUtils.escapeUtf8.