Fix for Javadoc escapeUtf8 in org.freeplane.core.util.TextUtils
In merged pull request #117 the new method `escapeUtf8` was added to
`org.freeplane.core.util.TextUtils`. There is a small mistake in the
example in Javadoc. It reads:

> input string: jalapeño
> output string: jalapeño

It should read

> input string: jalapeño
> output string: jalape\u00F1o

Somehow `\u00F1` gets unescaped when creating the Javadoc. This is
strange because the Javadoc preview shows it correct A double backslash
in the same text is not escaped, but it is enclosed with `<code>` tags.
I have enclosed `\u00F1` also with `<code>` tags. Hopefully that will
solve the problem.