use toUriFriendlyDecodedString() as old input displayed for editing by User
because it is more human-readable;
the string is URI-constructor friendly, so it can be encoded back to URI.
In SHTMLEditLinkAction, make sure href is encoded, although decoded input URI string is provided. It's achieved by adding `keepOriginalString` (true by default) to LinkController#createHyperlink, and using `false` in a call from SHTMLEditLinkAction.
In contrast, continue to use `true` for `keepOriginalString` in SetLinkByTextFieldAction and AttributePopupMenu, because based on manual tests, it has no negative impact