make sure #at's `reference` is decoded
MapExplorerController#getNodeAt treats `reference` literally, e.g. `#at(/'a%20b')` will be understood as "find a child of root with text `a%20b`" rather than the intended "with text `a b`".
`reference` is the _fragment_ part of URI, and until this change, it was extracted with URI#toString, which returns an encoded string. With this change, `reference` passed to MapExplorerController#getNodeAt is decoded as much as possible, via URI#getFragment.
Note: some characters aren't decoded by URI#getFragment, in particular `[` and `]` stay as `%5B` and `%5D`